requirements.txt
python-dotenv
deepgram-sdk==3.1.4
langchain-pinecone 
pymongo 
Flask-PyMongo
pytz
flask-cors
python-dotenv
validators
gunicorn
boto3
pyyaml
pinecone-client
python-docx
PyPDF2 
flask-sock
assemblyai
python-dotenv
ngrok
twilio
gevent

langgraph
langchain-community
langchain-anthropic
tavily-python
pandas
openai
langchain-groq
